+++ SampleTextFiles/tre.txt
@@ -1,0 +1,2 @@
+00001r1s001f1w00001L1E1A01P1²1'1I1W1D1"01³1/1y1i1e01t1n0001.01T1¹001J1V1S1z1-1:01j1C1d1l01m1g1o01c1h1 01a001p01b1v1,01Â1
+1u
