+++ SampleTextFiles/words200/file1.txt
@@ -1,0 +1,8 @@
+World peace is the concept of an ideal state of peace within and among all people and nations on Planet Earth¹. Different cultures, religions, philosophies, and organizations have varying concepts on how such a state would come about. Some of the possible ways to achieve world peace are:
+
+- Capitalist peace theory: This theory argues that free trade and economic interdependence reduce the incentives for war¹.
+- Democratic peace theory: This theory claims that democracies rarely or never wage war against each other because they share common values and respect human rights¹.
+- Cobdenism: This theory proposes that by removing tariffs and creating international free trade, wars would become impossible because nations would become dependent on each other¹.
+- Culture of peace: This is a set of values, attitudes, and behaviors that reject violence and promote dialogue, cooperation, and justice². Various religious and secular organizations have the stated aim of achieving world peace through addressing human rights, technology, education, engineering, medicine, or diplomacy¹.
+
+World peace is not only a desirable goal but also a necessary one for the survival and well-being of humanity. As John Lennon sang, "I hope someday you'll join us / And the world will be as one" ³.
