+++ SampleTextFiles/words200/file3.txt
@@ -1,0 +1,7 @@
+Video games are a popular form of entertainment for many children, but they also have positive and negative effects on their development. Depending on the type and content of the games, they can enhance or impair various skills and behaviors.
+
+Some of the positive effects of video games are that they can provide children with problem-solving skills, creativity, history and culture knowledge, social interaction, leadership skills, motivation, and memory improvement¹. Educational video games can also help children learn coding, literacy, math, and other subjects³. Playing video games with parents can enhance the learning outcomes and the parent-child bond³. Exergames, which involve physical movement and exercise, can improve children's decision-making and brain functions³.
+
+Some of the negative effects of video games are that they can cause health problems such as obesity, sleep disturbances, headaches, dry eyes, carpal tunnel syndrome, and muscular and postural disorders⁵. Violent and entertaining video games can also increase aggressive thoughts, feelings, and behaviors in children²⁴. They can also reduce children's interest in other activities and cause them to lose relationship, educational, or career opportunities². Playing video games for too long or too often can also lead to addiction and withdrawal symptoms².
+
+Therefore, video games have both good and bad effects on children. Parents should monitor their children's video game use and encourage them to play educational and non-violent games in moderation.
