+++ SampleTextFiles/words200/tre.txt
@@ -1,0 +1,2 @@
+0000rs00fw0000LEA0P²'IWD"0³/yie0tn000.0T¹00JVSz-:0jCdl0mgo0ch 0a00p0bv,0Â
+u
